# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SWAN GMAO API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workorders:
    get:
      operationId: getWorkOrders
      tags: [workorders]
      summary: List all work orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of work orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkOrder"
    post:
      operationId: createWorkOrder
      tags: [workorders]
      summary: Create a new work order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkOrderBody"
      responses:
        "201":
          description: Created work order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
  /workorders/{id}:
    get:
      operationId: getWorkOrder
      tags: [workorders]
      summary: Get work order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Work order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
    put:
      operationId: updateWorkOrder
      tags: [workorders]
      summary: Update a work order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkOrderBody"
      responses:
        "200":
          description: Updated work order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
    delete:
      operationId: deleteWorkOrder
      tags: [workorders]
      summary: Delete a work order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /zones:
    get:
      operationId: getZones
      tags: [zones]
      summary: List all zones
      parameters:
        - name: siteId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of zones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Zone"
    post:
      operationId: createZone
      tags: [zones]
      summary: Create a zone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateZoneBody"
      responses:
        "201":
          description: Created zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
  /zones/{id}:
    put:
      operationId: updateZone
      tags: [zones]
      summary: Update a zone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateZoneBody"
      responses:
        "200":
          description: Updated zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
    delete:
      operationId: deleteZone
      tags: [zones]
      summary: Delete a zone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    WorkOrder:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [corrective, preventive, predictive, inspection]
        priority:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [open, in_progress, completed, cancelled, on_hold]
        assetId:
          type: integer
        assetName:
          type: string
        technicianId:
          type: integer
        technicianName:
          type: string
        estimatedHours:
          type: number
        actualHours:
          type: number
        scheduledDate:
          type: string
        completedDate:
          type: string
        siteId:
          type: integer
        siteName:
          type: string
        zoneId:
          type: integer
        zoneName:
          type: string
        assignmentMode:
          type: string
          enum: [by_technician, by_zone, by_machine, by_type]
        createdAt:
          type: string
      required:
        - id
        - title
        - type
        - priority
        - status
        - createdAt
    CreateWorkOrderBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [corrective, preventive, predictive, inspection]
        priority:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [open, in_progress, completed, cancelled, on_hold]
        assetId:
          type: integer
        technicianId:
          type: integer
        estimatedHours:
          type: number
        scheduledDate:
          type: string
        siteId:
          type: integer
        zoneId:
          type: integer
        assignmentMode:
          type: string
          enum: [by_technician, by_zone, by_machine, by_type]
      required:
        - title
        - type
        - priority
...
    Zone:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        siteId:
          type: integer
        siteName:
          type: string
        description:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - createdAt
    CreateZoneBody:
      type: object
      properties:
        name:
          type: string
        siteId:
          type: integer
        description:
          type: string
      required:
        - name
...